requirements_3.8.txt
fsspec==2024.10.0
s3fs==2024.10.0
tifffile==2023.7.10
pillow==10.4.0
imagecodecs==2023.3.16
tqdm==4.67.1
matplotlib==3.7.5
requests==2.32.3
svgpathtools==1.6.1
svg.path
shapely==2.0.7